# openapi.yaml
openapi: 3.1.0
info:
  title: Adminly API
  version: 1.0.0
  description: >-
    Self-contained contract for the starter. Route Handlers under src/app/api
    implement it against an in-memory store. Swap `servers.url` + the generated
    schema for your real backend (Go/Huma, NestJS+Swagger, ...).
servers:
  - url: /api
paths:
  /users:
    get:
      operationId: listUsers
      summary: List users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUser"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
  /users/{id}:
    delete:
      operationId: deleteUser
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /items:
    get:
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: per_page, in: query, schema: { type: integer } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          content: { application/json: { schema: { $ref: "#/components/schemas/ItemList" } } }
    post:
      requestBody:
        content: { application/json: { schema: { $ref: "#/components/schemas/NewItem" } } }
      responses:
        "201":
          content: { application/json: { schema: { $ref: "#/components/schemas/Item" } } }
  /items/{id}:
    get:
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Item" } } } }
    put:
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody:
        content: { application/json: { schema: { $ref: "#/components/schemas/NewItem" } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Item" } } } }
    delete:
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "204": { description: No Content } }
  /items/bulk-delete:
    post:
      requestBody:
        content: { application/json: { schema: { type: object, properties: { ids: { type: array, items: { type: string } } } } } }
      responses: { "204": { description: No Content } }
  /items/options:
    get:
      parameters: [ { name: q, in: query, schema: { type: string } } ]
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Option" } } } } }
components:
  schemas:
    User:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        status:
          type: string
          enum: [active, invited, suspended]
        createdAt: { type: string }
    NewUser:
      type: object
      required: [name, email, role]
      properties:
        name: { type: string }
        email: { type: string }
        role: { type: string }
    Item:
      type: object
      required: [id, nama]
      properties:
        id: { type: string }
        nama: { type: string }
    NewItem:
      type: object
      required: [nama]
      properties:
        nama: { type: string }
    ItemList:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Item" } }
        meta:
          type: object
          required: [total, page, per_page]
          properties:
            total: { type: integer }
            page: { type: integer }
            per_page: { type: integer }
    Option:
      type: object
      required: [value, label]
      properties:
        value: { type: string }
        label: { type: string }